fix totalVoxels and polynom

+++ pipeline.py
@@ -4,6 +4,7 @@
 # ### Required libraries
 
 # In[21]:
+from __future__ import division
 import matplotlib.pyplot as plt
 import csv
 import pandas as pd
@@ -160,13 +161,13 @@ def Finn_preprocess(fmriFile):
         # get some info
         img = nib.load(fmriFile)
         hdr = img.header.structarr
-        nTRs = hdr['dim'][4]
+        nTRs = long(hdr['dim'][4])
         # retrieve TR
         TR = hdr['pixdim'][4]
         # retrieve dimensions
-        dim1 = hdr['dim'][1]
-        dim2 = hdr['dim'][2]
-        dim3 = hdr['dim'][3]
+        dim1 = long(hdr['dim'][1])
+        dim2 = long(hdr['dim'][2])
+        dim3 = long(hdr['dim'][3])
         ## DO PREPROCESSING:
         ## 1) Regress temporal drift from CSF and white matter (3rd order polynomial)
         print 'Step 1 (detrend WMCSF voxels, polynomial order 3)'
